Adjusted the input to be full-range
Fixed the wbpwmaudio call to also depend upon the same DEFAULT_RELOAD
value (sample rate).  wbpwmaudio now displays the expected aliasing
products.

Adjusted the test so that it starts with the first test, but also so that the
first test is the swept tone.  Pushed the two second pause until after the
two seconds of 440Hz.

Added some more commentary.